Defined storage on building level | updates states, control, output sets
I have defined the storage at the building level by way of defining the name of the storage states (e.g. "_tank_temperature" for the "building_scenario['building_name'] instead of self.building_zones['zone_name'] that was before.

I have modified the schema also.

ERROR: "IndexError: index -1 is out of bounds for axis 0 with size 0" @ line 2508.
The errors seems to be related once again to the lack of a constant for storage when it comes to disturbances.